Improve docker installation script for ubuntu setup.

+++ setup-ubuntu.sh
@@ -10,11 +10,18 @@ error() { echo -e "\033[1;31m[ERROR]\033[0m $1"; exit 1; }
 # ----------------------------------- SETUP -----------------------------------
 
 # --- apt setup
+log "# --- setting up package dependencies..."
+
 sudo apt update && sudo apt upgrade -y
 
 ### install package dependencies
 
-missing_packages=()
+# Function to check if a package is installed
+is_package_installed() {
+    dpkg -s "$1" &>/dev/null
+}
+
+MISSING_PACKAGES=()
 
 while IFS= read -r pkg; do
     # Remove inline comments and trim whitespace
@@ -24,20 +31,21 @@ while IFS= read -r pkg; do
     [[ -z "$pkg" ]] && continue
 
     # Check if the package is NOT installed
-    if ! dpkg -s "$pkg" &>/dev/null; then
-        missing_packages+=("$pkg")
+    if ! is_package_installed "$pkg"; then
+        MISSING_PACKAGES+=("$pkg")
     fi
 done < apt.pkg
 
 # Check if there are missing packages
-if [ "${#missing_packages[@]}" -gt 0 ]; then
-    log "Installing missing packages: ${missing_packages[@]}"
-    sudo apt install -y "${missing_packages[@]}"
+if [ "${#MISSING_PACKAGES[@]}" -gt 0 ]; then
+    log "Installing missing packages: ${MISSING_PACKAGES[@]}"
+    sudo apt install -y "${MISSING_PACKAGES[@]}"
 else
     log "All package dependencies are already installed."
 fi
 
 # --- ZSH setup
+log "# --- setting up ZSH..."
 
 ### install OhMyZSH!
 if [[ -d "$HOME/.oh-my-zsh" ]]; then
@@ -66,6 +74,8 @@ install_zsh_plugin "https://github.com/romkatv/powerlevel10k.git" "$ZSH_CUSTOM/t
 install_zsh_plugin "https://github.com/zsh-users/zsh-autosuggestions" "$ZSH_CUSTOM/plugins/zsh-autosuggestions"
 
 # --- install nerd fonts
+log "# --- installing nerd fonts..."
+
 FONT_DIR="${HOME}/.fonts"
 FONT_NAME="Mononoki Nerd"
 FONT_SRC_DIR="$PWD/fonts/${FONT_NAME}"
@@ -104,6 +114,8 @@ fi
 log "Font installation complete."
 
 # --- install asdf
+log "# --- installing asdf..."
+
 if [[ ! -d "$HOME/.asdf" ]]; then
     log "Installing asdf..."
     git clone https://github.com/asdf-vm/asdf.git "$HOME/.asdf" --branch v0.14.1 || error "Failed to install asdf."
@@ -114,29 +126,36 @@ fi
 # ----------------------------------- CONFIGS -----------------------------------
 
 create_symlink() {
-    src=$1
-    dest=$2
-    ln -sf "$src" "$dest" && log "Created symlink: $src -> $dest"
+    local src=$1
+    local dest=$2
+
+    if [[ -L "$dest" && "$(readlink "$dest")" == "$src" ]]; then
+        log "Symlink already exists: $dest -> $src. Skipping."
+    else
+        ln -sf "$src" "$dest" && log "Created symlink: $src -> $dest"
+    fi
 }
 
 # --- setup symlinks to config files
+log "# --- setting up symlinks for config files..."
 
 if ! [[ -d ${HOME}/.config/helix/themes ]]; then
     mkdir ${HOME}/.config/helix/themes
 fi
 
-create_symlink $PWD/.bashrc ${HOME}
-create_symlink $PWD/.zshrc ${HOME}
-create_symlink $PWD/.gitconfig ${HOME}
-create_symlink $PWD/.p10k.zsh ${HOME}
-create_symlink $PWD/helix/config.toml ${HOME}/.config/helix/
-create_symlink $PWD/helix/languages.toml ${HOME}/.config/helix/
-create_symlink $PWD/helix/themes/night_owl.toml ${HOME}/.config/helix/themes/
-create_symlink $PWD/tilda/config_0 ${HOME}/.config/tilda/
+create_symlink $PWD/.bashrc $HOME/.bashrc
+create_symlink $PWD/.zshrc $HOME/.zshrc
+create_symlink $PWD/.gitconfig $HOME/.gitconfig
+create_symlink $PWD/.p10k.zsh $HOME/.p10k.zsh
+create_symlink $PWD/helix/config.toml $HOME/.config/helix/config.toml
+create_symlink $PWD/helix/languages.toml $HOME/.config/helix/languages.toml
+create_symlink $PWD/helix/themes/night_owl.toml $HOME/.config/helix/themes/night_owl.toml
+create_symlink $PWD/tilda/config_0 $HOME/.config/tilda/config_0
 
 # ----------------------------------- INSTALLATION -----------------------------------
 
 # --- install asdf plugins and tools
+log "# --- installing asdf plugins and tools..."
 
 declare -A tools=(
     [python]=3.12.8
@@ -181,6 +200,7 @@ done
 log "All tools have been successfully installed and configured."
 
 # --- install helix language servers
+log "# --- installing helix language servers..."
 
 ### Ruff installation
 if command -v pip >/dev/null; then
@@ -209,6 +229,7 @@ else
 fi
 
 # --- install dev utils
+log "# --- installing dev utils..."
 
 # Ensure Cargo is available
 if ! command -v cargo >/dev/null; then
@@ -242,35 +263,75 @@ done
 log "Development utilities setup complete."
 
 # --- install Docker
+log "# --- installing docker..."
 
 log "Checking for Docker installation..."
+
+
+# Docker packages to check and install
+DOCKER_PACKAGES=(
+    docker-ce
+    docker-ce-cli
+    containerd.io
+    docker-buildx-plugin
+    docker-compose-plugin
+)
 
 if command -v docker >/dev/null 2>&1; then
     log "Docker is already installed. Skipping installation."
 else
     log "Docker is not installed. Proceeding with installation..."
 
-    # Set up Docker's official GPG key and repository
-    sudo install -m 0755 -d /etc/apt/keyrings || error "Failed to create keyrings directory."
-    sudo curl -fsSL https://download.docker.com/linux/ubuntu/gpg -o /etc/apt/keyrings/docker.asc || error "Failed to download Docker GPG key."
-    sudo chmod a+r /etc/apt/keyrings/docker.asc || error "Failed to set permissions on Docker GPG key."
+    if [[ ! -f /etc/apt/keyrings/docker.asc ]]; then
+        log "Adding Docker's GPG key..."
+        
+        # Set up Docker's official GPG key and repository
+        sudo install -m 0755 -d /etc/apt/keyrings || error "Failed to create keyrings directory."
+        sudo curl -fsSL https://download.docker.com/linux/ubuntu/gpg -o /etc/apt/keyrings/docker.asc || error "Failed to download Docker GPG key."
+        sudo chmod a+r /etc/apt/keyrings/docker.asc || error "Failed to set permissions on Docker GPG key."
 
-    log "Adding Docker repository to APT sources..."
-    echo \
-        "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/ubuntu \
-        $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | \
-        sudo tee /etc/apt/sources.list.d/docker.list > /dev/null || error "Failed to add Docker repository."
+        log "Adding Docker repository to APT sources..."
+        echo \
+            "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/ubuntu \
+            $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | \
+            sudo tee /etc/apt/sources.list.d/docker.list > /dev/null || error "Failed to add Docker repository."
+
+        log "Docker's GPG key and repository added successfully."
+    else
+        log "Docker's GPG key and repository already configured. Skipping."
+    fi
 
     log "Updating APT package index..."
-    sudo apt-get update || error "Failed to update APT package index."
+    sudo apt update || error "Failed to update APT package index."
 
-    log "Installing Docker packages..."
-    sudo apt-get install -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin || error "Failed to install Docker packages."
+    MISSING_PACKAGES=()
+    for pkg in "${DOCKER_PACKAGES[@]}"; do
+        if ! is_package_installed "$pkg"; then
+            MISSING_PACKAGES+=("$pkg")
+        fi
+    done
 
-    log "Docker installation completed successfully."
+    if [[ ${#MISSING_PACKAGES[@]} -gt 0 ]]; then
+        log "Installing missing Docker packages: ${MISSING_PACKAGES[*]}..."
+        sudo apt install -y "${MISSING_PACKAGES[@]}" || error "Failed to install Docker packages."
+    else
+        log "All required Docker packages are already installed."
+    fi
+
 fi
 
+# Check and add user to the Docker group if not already a member
+if groups "$USER" | grep -q "\bdocker\b"; then
+    log "User '$USER' is already a member of the 'docker' group. Skipping."
+else
+    log "Adding user '$USER' to the 'docker' group..."
+    sudo usermod -aG docker "$USER" || error "Failed to add user '$USER' to the 'docker' group."
+    log "User '$USER' added to the 'docker' group. You may need to re-login for the changes to take effect."
+fi    
+
+log "Docker installation completed successfully."
 # --- install pgAdmin4
+log "# --- installing pgAdmin4"
 
 log "Checking for pgAdmin4 installation..."
